+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
  - Add temperature compensation option for the Atlas Scientific Electrical Conductivity and Dissolved Oxygen Inputs
+ - Add Inputs: Atlas Scientific Flow Sensor, Atlas Scientific RGB Color Sensor
 
 
 ## 8.3.0 (2020-02-21)